requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.1.7
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jwcrypto==1.4.2
numpy==1.22.3
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
requests==2.30.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.2
